tracingpolicy: add tags field to categorize events
In the process of producing user friendly events, we have added in
1842f69ff28c6ca "message" field so user can understand why this event
was produced instead of cryptic message generated from code.

Now to categorize events we need a "tags" field that is a list of tags.

The Sigma Tags Specification is a good starting point:
https://github.com/SigmaHQ/sigma-specification/blob/main/Tags_specification.md

We will first define namespaces for events, and we start with the
"observability" namespace.

observability: is the ability to collect and export data about the internal
system state.
- "observability.filesystem": this event is about filesystem operations.

- "observability.privilege": this event is about the rights and permissions
    granted to a process or a user.

- "observability.process": this event is about an instance of a Linux program
    that is being executed.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>