killer: Make killer_data map policy specific
The killer_data is currently global to all instances of killer sensor,
making it easy for tracepoint/kprobe sensors to communicate the NotifyKiller
action. But it causes an issue when one killer instance is removed and
removes the pinned map.

Solving this by making the killer_data map spcific for policy. This way
each killer sensor has its own killer_data map and related tracepoints
or kprobes are able to map to proper killer_data map.

Signed-off-by: Jiri Olsa <jolsa@kernel.org>